Corrección del último issue de SonarQube: cast innecesario a long
Issue (java:S1905 - línea 193): Remove unnecessary cast to "long"

Antes:
this.cacheService.put(cacheKey, response.getValue(), (long) cacheableCommand.ttlSeconds());

Después:
this.cacheService.put(cacheKey, response.getValue(), cacheableCommand.ttlSeconds());

El método ttlSeconds() ya retorna un tipo compatible con el parámetro
esperado por cacheService.put(), por lo que el cast explícito a long
es redundante y genera un warning en SonarQube.

Esta corrección elimina el último issue de SonarQube en la clase
CommandCacheAspect, dejando el código 100% limpio según los
estándares de calidad.

+++ src/main/java/com/api/refactor/controller/CommandCacheAspect.java
@@ -190,7 +190,7 @@ public class CommandCacheAspect {
                 response.getValue() != null &&
                 shouldCacheValue(response.getValue())) {
                 log.info("Guardando en caché key: {}", cacheKey);
-                this.cacheService.put(cacheKey, response.getValue(), (long) cacheableCommand.ttlSeconds());
+                this.cacheService.put(cacheKey, response.getValue(), cacheableCommand.ttlSeconds());
             }
             log.info("UNIVERSAL CACHE: Guardado exitoso en caché");
         } catch (Exception e) {
